+++ package/src/scripts/methods/changeTheme.ts
